Fix static analysis low issues
:Release Notes:
Fix static analysis low issues

:Detailed Notes:
Fix static analysis low issues

:Testing Performed:
Local Test

:QA Notes:
N/A

:Issues Addressed:
[WRP-26934] Fix static analysis issues

Reviewed-on: http://gpro.lge.com/c/webosose/edge-ai-computer-vision/+/371872
Package-Only-Build: Seungchul Han <seungchul.han@lge.com>
Package-Only-Build-LGSI: Seungchul Han <seungchul.han@lge.com>
Tested-by: Seungchul Han <seungchul.han@lge.com>

+++ include/aif/text/TextDescriptor.h
@@ -23,10 +23,10 @@ public:
     bool addBboxes(const std::vector<cv::Rect>& bboxes);
     bool addBoxes(const std::vector<std::vector<cv::Point>>& boxes);
 
-    int getTextRectCount() const { return m_textRectCount; }
+    size_t getTextRectCount() const { return m_textRectCount; }
 
 private:
-    int m_textRectCount;
+    size_t m_textRectCount;
 };
 
 } // end of namespace aif
